pyproject.toml
[tool.poetry]
name = "fastapi-docencia"
version = "2.0.0"
description = "backend docencia"
authors = ["chaknogf <chaknogf@gmail.com>"]
license = "gpl"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-jose = "3.5.0"
aiosmtplib = "4.0.2"
annotated-types = "0.7.0"
anyio = "4.11.0"
async-timeout = "5.0.1"
babel = "2.17.0"
backrefs = "5.9"
bcrypt = "5.0.0"
blinker = "1.9.0"
cachetools = "6.2.1"
certifi = "2025.11.12"
cffi = "2.0.0"
charset-normalizer = "3.4.4"
click = "8.3.1"
colorama = "0.4.6"
contourpy = "1.3.1"
cryptography = "46.0.3"
cycler = "0.12.1"
dnspython = "2.8.0"
ecdsa = "0.19.1"
email-validator = "2.3.0"
et-xmlfile = "2.0.0"
fakeredis = "2.32.0"
fastapi = "0.122.0"
fastapi-mail = "1.5.8"
flake8 = "7.3.0"
fonttools = "4.56.0"
ghp-import = "2.1.0"
google-auth = "2.41.1"
greenlet = "3.2.1"
h11 = "0.16.0"
idna = "3.11"
isort = "7.0.0"
jinja2 = "3.1.6"
kiwisolver = "1.4.8"
markdown = "3.9"
markupsafe = "3.0.3"
matplotlib = "3.10.1"
mccabe = "0.7.0"
mergedeep = "1.3.4"
mkdocs = "1.6.1"
mkdocs-get-deps = "0.2.0"
mkdocs-material = "9.6.22"
mkdocs-material-extensions = "1.3.1"
narwhals = "1.30.0"
numpy = "2.2.3"
openpyxl = "3.1.5"
packaging = "25.0"
paginate = "0.5.7"
pandas = "2.2.3"
passlib = "1.7.4"
pathspec = "0.12.1"
pillow = "11.1.0"
pipdeptree = "2.30.0"
platformdirs = "4.5.0"
plotly = "6.0.0"
psycopg2-binary = "2.9.11"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.2"
pycodestyle = "2.14.0"
pycparser = "2.23"
pydantic = "2.12.5"
pydantic-settings = "2.12.0"
pydantic-core = "2.41.5"
pyflakes = "3.4.0"
pygments = "2.19.2"
pyjwt = "2.10.1"
pymdown-extensions = "10.16.1"
pymysql = "1.1.1"
pyparsing = "3.2.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.2.1"
python-multipart = "0.0.20"
pytz = "2025.1"
pyyaml = "6.0.3"
pyyaml-env-tag = "1.1"
redis = "7.0.1"
regex = "2025.11.3"
requests = "2.32.5"
rsa = "4.9.1"
six = "1.17.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
sqlalchemy = "2.0.44"
starlette = "0.50.0"
typing-inspection = "0.4.2"
typing-extensions = "4.15.0"
tzdata = "2025.1"
urllib3 = "2.5.0"
uvicorn = "0.38.0"
watchdog = "6.0.0"
annotated-doc = "0.0.4"
argon2-cffi = "25.1.0"
argon2-cffi-bindings = "25.1.0"
httpcore = "1.0.9"
httpx = "0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
